+++ ios/Classes/SwiftMuxRtmpFlutterPlugin.swift
@@ -6,9 +6,12 @@ public class SwiftMuxRtmpFlutterPlugin: NSObject, FlutterPlugin {
     let channel = FlutterMethodChannel(name: "mux_rtmp_flutter", binaryMessenger: registrar.messenger())
     let instance = SwiftMuxRtmpFlutterPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
+
+    let factory = MuxRtmpViewFactory(messenger: registrar.messenger())
+    registrar.register(factory, withId: "MuxRtmpView")
   }
 
-  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+  public func handle(_: FlutterMethodCall, result: @escaping FlutterResult) {
     result("iOS " + UIDevice.current.systemVersion)
   }
 }